+++ example/indexing-right.m
@@ -78,10 +78,10 @@ H=[1:7;2:8]'
 [1,2;3,4](1,2) % literal indexing
 % test special indexing ('end' and collon (:))
 clear
-A = [1,2,3; 4,5,6; 7,8,9]
-A(2,3)
+A = [1,2,3,4,5; 6,7,8,9,10; 11,12,13,14,15]
+A(2,5)
 A(2,end)
-A(3,2)
+A(3,4)
 A(3,end-1)
 B = [16,5,9,4,2,11,7,14]
 B([5:8, 1:4]) % Extract and swap the halves of B